requirements.txt
pdf2image==1.5.4
pillow==6.0.0
BotMe==0.2